+++ apps/extension/src/ui/domains/Staking/Bittensor/hooks/useBittensorBondWizard.ts
@@ -5,7 +5,7 @@ import {
   subNativeTokenId,
   TokenId,
 } from "@talismn/chaindata-provider"
-import { Address } from "extension-core"
+import { Address, isAccountOfType } from "extension-core"
 import { useCallback, useEffect, useMemo, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { BehaviorSubject } from "rxjs"
@@ -15,14 +15,7 @@ import { Hex } from "viem"
 import { provideContext } from "@talisman/util/provideContext"
 import { useScaleApi } from "@ui/hooks/sapi/useScaleApi"
 import { useAnalytics } from "@ui/hooks/useAnalytics"
-import {
-  useAccountByAddress,
-  useAppState,
-  useFeatureFlag,
-  usePortfolioBalances,
-  useToken,
-  useTokenRates,
-} from "@ui/state"
+import { useAccountByAddress, usePortfolioBalances, useToken, useTokenRates } from "@ui/state"
 
 import { useExistentialDeposit } from "../../../../hooks/useExistentialDeposit"
 import { useFeeToken } from "../../../SendFunds/useFeeToken"
@@ -117,8 +110,6 @@ const useDtaoToken = (networkId: string, netuid: number, hotkey?: string) => {
 
 const useBittensorBondWizardProvider = () => {
   const { t } = useTranslation()
-  const isSeekTaoDiscountEnabled = useFeatureFlag("SEEK_TAO_DISCOUNT")
-  const [hideSeekDiscountDrawer] = useAppState("hideSeekTaoDiscountDrawer")
   const { genericEvent } = useAnalytics()
   const { allBalances } = usePortfolioBalances()
 
@@ -139,6 +130,7 @@ const useBittensorBondWizardProvider = () => {
   ] = useState(() => wizardOpenState$.getValue())
   const nativeTokenId = useMemo(() => (networkId ? subNativeTokenId(networkId) : null), [networkId])
   const dtaoToken = useDtaoToken(networkId ?? "", netuid ?? 0, hotkey ?? undefined)
+  const [isMevProtectionEnabled, setIsMevProtectionEnabled] = useState(false)
 
   const dtaoBalance = useBalance(allBalances, address, dtaoToken?.id)
   const nativeBalance = useBalance(allBalances, address, nativeTokenId)
@@ -154,6 +146,17 @@ const useBittensorBondWizardProvider = () => {
   const seekDiscountDrawer = useOpenClose()
 
   const { data: sapi } = useScaleApi(nativeToken?.networkId)
+
+  const isMevShieldDisabled = useMemo(() => {
+    // no need for root staking
+    // supported only for hot wallets
+    return !netuid || !isAccountOfType(account, "keypair")
+  }, [netuid, account])
+
+  const withMevShield = useMemo(
+    () => !isMevShieldDisabled && isMevProtectionEnabled,
+    [isMevShieldDisabled, isMevProtectionEnabled],
+  )
 
   const {
     alphaPrice,
@@ -321,21 +324,13 @@ const useBittensorBondWizardProvider = () => {
     if (stakeDirection === "unbond") {
       return totalStakedPlancks
     }
-    if (!nativeBalance || !existentialDeposit || !feeEstimate || typeof talismanFee !== "bigint")
-      return null
+    if (!nativeBalance || !existentialDeposit || !feeEstimate) return null
     if (existentialDeposit.planck + feeEstimate * 11n > nativeBalance.transferable.planck)
       return null
     const maxRootStake =
       nativeBalance.transferable.planck - existentialDeposit.planck - feeEstimate * 11n
     return maxRootStake
-  }, [
-    stakeDirection,
-    nativeBalance,
-    existentialDeposit,
-    feeEstimate,
-    talismanFee,
-    totalStakedPlancks,
-  ])
+  }, [stakeDirection, nativeBalance, existentialDeposit, feeEstimate, totalStakedPlancks])
 
   const newStakeTotal = useMemo(() => {
     if (stakeDirection === "unbond") {
@@ -482,12 +477,6 @@ const useBittensorBondWizardProvider = () => {
   }, [stakeDirection, position, setStep, step])
 
   useEffect(() => {
-    // if subnet staking, open seek discount drawer if it has not been displayed yet
-    if (isSeekTaoDiscountEnabled && !hideSeekDiscountDrawer && stakeType === "subnet")
-      seekDiscountDrawer.open()
-  }, [hideSeekDiscountDrawer, isSeekTaoDiscountEnabled, seekDiscountDrawer, stakeType])
-
-  useEffect(() => {
     // on mount, if stake type is not set, display the stake type select drawer
     if (!stakeType && stakeDirection === "bond") stakeTypeDrawer.open()
   }, [stakeType, stakeDirection, stakeTypeDrawer])
@@ -534,6 +523,9 @@ const useBittensorBondWizardProvider = () => {
     talismanFee,
     amountOut,
     priceImpact,
+    withMevShield,
+    isMevShieldDisabled,
+    setIsMevProtectionEnabled,
     setAddress,
     setNetuid,
     setHotkey,