+++ apps/extension/src/ui/domains/Swap/swap-modules/stealthex-swap-module.ts
@@ -145,17 +145,17 @@ const supportedEvmChains: Record<string, ViemChain | undefined> = {
  */
 const specialAssets: Record<string, Omit<SwappableAssetBaseType, "context">> = {
   "mainnet::dot": {
-    id: subNativeTokenId("polkadot"),
-    name: "Polkadot",
+    id: subNativeTokenId("polkadot-asset-hub"),
+    name: "Polkadot Asset Hub",
     symbol: "DOT",
-    chainId: "polkadot",
+    chainId: "polkadot-asset-hub",
     networkType: "substrate",
   },
   "polkadot::ksm": {
-    id: subNativeTokenId("kusama"),
-    name: "Kusama",
+    id: subNativeTokenId("kusama-asset-hub"),
+    name: "Kusama Asset Hub",
     symbol: "KSM",
-    chainId: "kusama",
+    chainId: "kusama-asset-hub",
     networkType: "substrate",
   },
   "polkadot::usdt": {