+++ packages/extension-core/src/domains/app/store.app.ts
@@ -39,7 +39,6 @@ export type AppStoreData = {
   hideManageAccountsWelcome?: boolean
   hideBittensorSubnetStakeWarning?: boolean
   hideGetStarted?: boolean
-  hideSeekTaoDiscountDrawer?: boolean
 
   // dismissed banners
   hideUnifiedAddressBanner?: boolean